web app temperature/requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
silhouette==0.0.1a0
sqlparse==0.5.1
